requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.23
django-extensions==3.2.3
djangorestframework==3.15.2
pygraphviz==1.11
sqlparse==0.5.3
typing-extensions==4.13.2
